Callback Test: Split float tests and test linux
Float tests only work on Linux with CGO_ENABLED=1 which was documented
